Implement switchable player characters
m_iPlayerCharacter is added to CBasePlayer, which is set from a field on
the player_info_start, which now links to CNtksStart instead of
CPointEntity.

A singleton PlayerCharacterStats is added to provide the different
character stats. Currently it merely supports swapping out viewmodels.
This is done by adding new keys for each character to the
weapon-scripts.